requirements.txt
Flask==2.2.2
pandas==1.5.3
Requests==2.31.0
Pillow==10.0.0
python-dotenv=latest
